clip: move model cgraphs into their own files
